Refactor fixtures.response to improve pytest introspection * Remove AssertRespIs class * Rename response.py -> test_response.py so it is included in test collection process
