+++ tests/account/test_api.py
@@ -359,7 +359,7 @@ class TestUpdateAPIKey:
 
 
 @pytest.mark.parametrize("error", [None, "404"])
-async def test_remove_api_key(error, spawn_client, assert_resp_is):
+async def test_remove_api_key(error, spawn_client, resp_is):
     client = await spawn_client(authorize=True)
 
     if not error:
@@ -375,14 +375,14 @@ async def test_remove_api_key(error, spawn_client, assert_resp_is):
     resp = await client.delete("/api/account/keys/foobar_0")
 
     if error:
-        await assert_resp_is.not_found(resp)
+        await resp_is.not_found(resp)
         return
 
     assert resp.status == 204
     assert await client.db.keys.count_documents({}) == 0
 
 
-async def test_remove_all_api_keys(spawn_client, assert_resp_is):
+async def test_remove_all_api_keys(spawn_client):
     client = await spawn_client(authorize=True)
 
     await client.db.keys.insert_many([
